Final Commit
Entered all the necessary variables and constructors for basic operation of the game. Found and debugged several interactions with buttons and players.

Cleaned Up css to make it look warmer.

There is a playerManager const that arises an error because it's never used but it's necessary for structure as it contains important data.

That's it!